+++ tests/test_domain.py
@@ -15,15 +15,16 @@ from tests.conftest import DEFAULT_DOMAIN_PATH
 
 def test_create_train_data_no_history(default_domain):
     featurizer = BinaryFeaturizer()
-    X, y = extract_training_data_from_file(
+    training_data = extract_training_data_from_file(
             "data/dsl_stories/stories_defaultdomain.md",
             augmentation_factor=0,
             domain=default_domain,
             featurizer=featurizer,
             max_history=1
     )
-    assert X.shape == (11, 1, 10)
-    decoded = [featurizer.decode(X[i, :, :], default_domain.input_features)
+    assert training_data.X.shape == (11, 1, 10)
+    decoded = [featurizer.decode(training_data.X[i, :, :],
+                                 default_domain.input_features)
                for i in range(0, 11)]
     assert decoded == [
         [None],
@@ -45,15 +46,16 @@ def test_create_train_data_no_history(default_domain):
 
 def test_create_train_data_with_history(default_domain):
     featurizer = BinaryFeaturizer()
-    X, y = extract_training_data_from_file(
+    training_data = extract_training_data_from_file(
             "data/dsl_stories/stories_defaultdomain.md",
             augmentation_factor=0,
             domain=default_domain,
             featurizer=featurizer,
             max_history=4
     )
-    assert X.shape == (11, 4, 10)
-    decoded = [featurizer.decode(X[i, :, :], default_domain.input_features)
+    assert training_data.X.shape == (11, 4, 10)
+    decoded = [featurizer.decode(training_data.X[i, :, :],
+                                 default_domain.input_features)
                for i in range(0, 11)]
     assert decoded == [
         [
