+++ tests/test_dsl.py
@@ -8,7 +8,7 @@ import io
 from rasa_core.events import SlotSet, ActionExecuted
 from rasa_core.training import extract_stories_from_file, \
     extract_story_graph_from_file
-from rasa_core.training.dsl import Story
+from rasa_core.training.structures import Story
 
 
 def test_can_read_test_story(default_domain):
