+++ tests/test_training.py
@@ -20,8 +20,8 @@ def test_story_visualization(default_domain):
     story_steps = StoryFileReader.read_from_file(
             "data/dsl_stories/stories.md", default_domain,
             interpreter=RegexInterpreter())
-    generated_graph = visualize_stories(story_steps)
-    assert len(generated_graph.nodes()) == 21
+    generated_graph = visualize_stories(story_steps, default_domain)
+    assert len(generated_graph.nodes()) == 20
 
 
 def test_training_script(tmpdir):
